override am__is_gnu_make to fix make check
There is a bug in automake
https://debbugs.gnu.org/cgi/bugreport.cgi?bug=18744
that causes make check to fail with 'Arg list too long'.
I applied the patch by overriding the problematic automake variable in
the Makefile.am.  The generation of Makefile.in does not override
user-specified variables, so the effect is as if I had applied the bug
report patch to automake itself.