avoid intel compiler bug?
Code such as the following caused the intel compiler to insert a value
of -1 in the second (1-th) position, and only for 8- and 16-bit versions.
This was verified for icc 15.0.1 and 18.0.0 using -O2 and -O3
optimization.

```C
__m256i vNegInfFront = _mm256_set_epi16(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,NEG_LIMIT);
```

Unfortunately, this was not reproducible in stand-alone test codes.
Only during the scan functions would this behavior occur.  The scan
functions were rewritten to avoid using the set function.