novec trace routines combine all tables
This begins the transition to using a single traceback table instead of
three separate tables.  It was quickest to demonstrate the change using
the novec routines; the vectorized routines use the old trace allocator
as well as the old cigar and traceback code.  The test_verify_cigar
program works for both old a new while during the interim.