windows fixes for parasail_aligner
stdin_has_data() was reporting false positives and then hanging during a false read of non-existent input.  Instead of using WaitForSingleObject(), the file length is read.  In some cases, fseek is attempted on stdin.  This seems to work.

Further, the kseq.h header is updated with the latest fixes from attractivechaos/klib.  This seems to fix additional bugs of parsing input files on windows.

https://github.com/attractivechaos/klib/blob/master/kseq.h